--- a/.gitattributes
+++ b/.gitattributes
@@ -1,0 +1,1 @@
+cldr.go linguist-generated=false