--- a/internal/gen/cldr.go.tmpl
+++ b/internal/gen/cldr.go.tmpl
@@ -30,7 +30,39 @@ var monthLookup = map[string]monthIndexes{
 {{- end }}
 }
 
-func fmtMonth(locale string, calendar calendarType, context, width string) func(int, string) string {
+// fmtYear formats year as numeric. The assumption is that f is "06" or "2006".
+func fmtYear(digits digits) func (y int, f string) string {
+	return func(y int, f string) string {
+		year := strconv.Itoa(y)
+
+		// ptime.Time.Format is very slow. Make it fast!
+		if f == "06" {
+			switch len(year) {
+			default:
+				year = year[len(year)-2:]
+			case 1:
+				year = "0" + year
+			case 0, 2: // noop, isSliceInBounds()
+			}
+		}
+
+		return digits.Sprint(year)
+	}
+}
+
+// fmtMonth formats month as numeric. The assumption is that f is "1" or "01".
+func fmtMonth(digits digits) func (m int, f string) string {
+	return func(m int, f string) string {
+		if f == "01" && m <= 9 {
+			return digits.Sprint("0" + strconv.Itoa(m))
+		}
+
+		return digits.Sprint(strconv.Itoa(m))
+	}
+}
+
+// fmtMonthName formats month as name.
+func fmtMonthName(locale string, calendar calendarType, context, width string) func(int, string) string {
   indexes := monthLookup[locale]
 
   i := calendar*6
@@ -128,20 +160,12 @@ func fmtMonth{{ title $calendarType }}(locale language.Tag, digits digits) func(
 {{ else }}
 func fmtMonth{{ title $calendarType }}(_ language.Tag, digits digits) func(month int, format string) string {
 {{ end }}
-  fmt := func(m int, f string) string {
-    if f == "01" && m <= 9 {
-      return digits.Sprint("0"+strconv.Itoa(m))
-    }
-
-    return digits.Sprint(strconv.Itoa(m))
-  }
-
 {{ if eq (len $formats.M.Fmt) 0 }}
-  return fmt
+  return fmtMonth(digits)
 {{ else }}
   switch lang.String() {
   default:
-    return fmt
+    return fmtMonth(digits)
   {{- range $fmt, $languages := $formats.M.Fmt }}
   case "{{ join $languages "\", \"" }}":
     return {{ $fmt }}
@@ -163,8 +187,6 @@ func fmtDay{{ title $calendarType }}(_ language.Tag, digits digits) func(day int
 
     return digits.Sprint(strconv.Itoa(d))
   }
-
-
 {{ if eq (len $formats.D.Fmt) 0 }}
   return fmt
 {{ else }}
@@ -193,7 +215,7 @@ func (f *gregorianDateTimeFormat) SetTime(v time.Time) {
 }
 
 func (f *gregorianDateTimeFormat) Year(format string) string {
-  return f.fmtYear(f.digits.Sprint(f.time.Format(format)))
+  return f.fmtYear(fmtYear(f.digits)(f.time.Year(), format))
 }
 
 func (f *gregorianDateTimeFormat) Month(format string) string {
@@ -218,20 +240,7 @@ func (f *persianDateTimeFormat) SetTime(v time.Time) {
 
 // Year returns formatted year (only "06" and "2006" is supported).
 func (f *persianDateTimeFormat) Year(format string) string {
-  year := strconv.Itoa(f.time.Year())
-
-  // ptime.Time.Format is very slow. Make it fast!
-  if format == "06" {
-    switch len(year) {
-    default:
-      year = year[len(year)-2:]
-    case 1:
-      year = "0" + year
-    case 0, 2: // noop, isSliceInBounds()
-    }
-  }
-
-  return f.fmtYear(f.digits.Sprint(year))
+  return f.fmtYear(fmtYear(f.digits)(f.time.Year(), format))
 }
 
 func (f *persianDateTimeFormat) Month(format string) string {
@@ -255,20 +264,7 @@ func (f *buddhistDateTimeFormat) SetTime(v time.Time) {
 }
 
 func (f *buddhistDateTimeFormat) Year(format string) string {
-  year := strconv.Itoa(f.time.Year())
-
-  // ptime.Time.Format is very slow. Make it fast!
-  if format == "06" {
-    switch len(year) {
-    default:
-      year = year[len(year)-2:]
-    case 1:
-      year = "0" + year
-    case 0, 2: // noop, isSliceInBounds()
-    }
-  }
-
-  return f.fmtYear(f.digits.Sprint(year))
+  return f.fmtYear(fmtYear(f.digits)(f.time.Year(), format))
 }
 
 func (f *buddhistDateTimeFormat) Month(format string) string {