--- a/internal/gen/cldr.go.tmpl
+++ b/internal/gen/cldr.go.tmpl
@@ -87,6 +87,17 @@ func fmtMonthName(locale string, calendar calendarType, context, width string) f
   }
 }
 
+// fmtDay formats day as numeric. The assumption is that f is "2" or "02".
+func fmtDay(digits digits) func(d int, f string) string {
+	return func(d int, f string) string {
+		if f == "02" && d <= 9 {
+			return digits.Sprint("0" + strconv.Itoa(d))
+		}
+
+		return digits.Sprint(strconv.Itoa(d))
+	}
+}
+
 type numberingSystem int
 
 const (
@@ -180,13 +191,7 @@ func fmtDay{{ title $calendarType }}(locale language.Tag, digits digits) func(da
 {{ else }}
 func fmtDay{{ title $calendarType }}(_ language.Tag, digits digits) func(day int, format string) string {
 {{ end }}
-  fmt := func(d int, f string) string {
-    if f == "02" && d <= 9 {
-      return digits.Sprint("0"+strconv.Itoa(d))
-    }
-
-    return digits.Sprint(strconv.Itoa(d))
-  }
+  fmt := fmtDay(digits)
 {{ if eq (len $formats.D.Fmt) 0 }}
   return fmt
 {{ else }}