```console
$ benchstat a.txt b.txt
goos: darwin
goarch: arm64
pkg: go.expect.digital/intl
cpu: Apple M2 Max
                         │    a.txt    │               b.txt                │
                         │   sec/op    │   sec/op     vs base               │
NewDateTime/fa-IR-12       207.5n ± 1%   207.8n ± 1%       ~ (p=0.382 n=10)
NewDateTime/lv-LV-12       257.4n ± 2%   257.1n ± 1%       ~ (p=0.516 n=10)
NewDateTime/dz-BT-12       260.0n ± 1%   261.4n ± 0%       ~ (p=0.271 n=10)
DateTime_Format/fa-IR-12   835.8n ± 1%   844.9n ± 1%  +1.08% (p=0.002 n=10)
DateTime_Format/lv-LV-12   190.2n ± 1%   173.3n ± 1%  -8.86% (p=0.000 n=10)
DateTime_Format/dz-BT-12   475.0n ± 1%   454.5n ± 0%  -4.33% (p=0.000 n=10)
geomean                    318.7n        312.4n       -1.98%

                         │   a.txt    │                b.txt                 │
                         │    B/op    │    B/op     vs base                  │
NewDateTime/fa-IR-12       269.0 ± 0%   269.0 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12       221.0 ± 0%   221.0 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12       221.0 ± 0%   221.0 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12   104.0 ± 0%   104.0 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   24.00 ± 0%   32.00 ± 0%  +33.33% (p=0.000 n=10)
DateTime_Format/dz-BT-12   144.0 ± 0%   144.0 ± 0%        ~ (p=1.000 n=10) ¹
geomean                    129.5        135.9        +4.91%
¹ all samples are equal

                         │   a.txt    │                b.txt                │
                         │ allocs/op  │ allocs/op   vs base                 │
NewDateTime/fa-IR-12       5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12       5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12       5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12   23.00 ± 0%   23.00 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   24.00 ± 0%   24.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    8.069        8.069       +0.00%
¹ all samples are equal
```